requirements.txt
# ElevenLabs Agent Client Dependencies
# =====================================

# HTTP requests with connection pooling and retry logic
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Type hints for older Python versions (optional)
typing-extensions>=4.8.0

# =============================================================================
# FastAPI Dependencies
# =============================================================================

# FastAPI web framework
fastapi>=0.109.0

# ASGI server for running FastAPI
uvicorn[standard]>=0.27.0

# Pydantic for data validation (FastAPI dependency)
pydantic>=2.5.0

# Settings management
pydantic-settings>=2.1.0
